Split out getSitePosts
